% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missSOM.R
\name{missSOM}
\alias{missSOM}
\title{missSOM}
\description{
The Self-Organizing Maps with Built-in Missing Data Imputation. Missing values are imputed and regularly updated during the online Kohonen algorithm. Our method can be used for data visualisation, clustering or imputation of missing data. It is an extension of the online algorithm of the kohonen package.
}
\details{
Self-Organizing Maps with Built-in Missing Data Imputation
}
\author{
you <youremail>
}
