% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nir.R
\docType{data}
\name{nir}
\alias{nir}
\title{Title  Near-infrared data with temperature effects}
\description{
A data object containing near-infrared spectra of ternary mixtures of ethanol, water and iso-propanol, measured at five different temperatures (30, 40, ..., 70 degrees Centigrade).
}
\references{
F. Wulfert , W.Th. Kok, A.K. Smilde: Anal. Chem. 1998, 1761-1767
}
\author{
My Name \email{blahblah@roxygen.org}
}
\keyword{dataset}
