% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wines.R
\docType{data}
\name{wines}
\alias{wines}
\alias{vintages}
\title{Wine data}
\source{
\url{http://kdd.ics.uci.edu}
}
\description{
A data frame containing 177 rows and thirteen columns; object 
\code{vintages} contains the class labels. 

These data are the results of chemical analyses of wines grown in the same region in Italy (Piedmont) but derived from three different cultivars: Nebbiolo, Barberas and Grignolino grapes. The 
wine from the Nebbiolo grape is called Barolo. The data contain the quantities of several constituents found in each of the three types of wines, as well as 
some spectroscopic variables.
}
\references{
M. Forina, C. Armanino, M. Castino and M. Ubigli. Vitis, 25:189-201 (1986)
}
\author{
My Name \email{blahblah@roxygen.org}
}
\keyword{dataset}
