% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast.R
\docType{data}
\name{yeast}
\alias{yeast}
\title{Title  Yeast cell-cycle data}
\description{
Microarray cell-cycle data for 800 yeast genes, arrested with six different methods, arranged in a list. Additional class information is present as well.
}
\references{
P. Spellman et al., Mol. Biol. Cell 9, 3273-3297 (1998)
}
\author{
My Name \email{blahblah@roxygen.org}
}
\keyword{dataset}
